refactor: add index.ts to components folder
